Reworked data generation scripts
